+++ ext/log-processor/src/test/java/core/log/service/ActionLogForwarderTest.java
@@ -2,7 +2,6 @@ package core.log.service;
 
 import core.framework.kafka.MessagePublisher;
 import core.framework.log.message.ActionLogMessage;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -11,6 +10,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 
 /**
  * @author neo
@@ -19,19 +19,50 @@ import static org.mockito.Mockito.verify;
 class ActionLogForwarderTest {
     @Mock
     MessagePublisher<ActionLogMessage> publisher;
-    private ActionLogForwarder forwarder;
-
-    @BeforeEach
-    void createActionLogForwarder() {
-        forwarder = new ActionLogForwarder(publisher, "action", List.of("website"), List.of());
-    }
 
     @Test
-    void forward() {
+    void forwardWithMatchedApp() {
+        var forwarder = new ActionLogForwarder(publisher, "action", List.of("website"), List.of(), List.of());
+
         var message = new ActionLogMessage();
         message.app = "website";
         message.traceLog = "trace";
         forwarder.forward(List.of(message));
         verify(publisher).publish("action", null, message);
+    }
+
+    @Test
+    void forwardWithMatchedResult() {
+        var forwarder = new ActionLogForwarder(publisher, "action", List.of("website"), List.of("OK"), List.of());
+
+        var message = new ActionLogMessage();
+        message.app = "website";
+        message.result = "OK";
+        message.traceLog = "trace";
+        forwarder.forward(List.of(message));
+        verify(publisher).publish("action", null, message);
+    }
+
+    @Test
+    void forwardWithMismatchedResult() {
+        var forwarder = new ActionLogForwarder(publisher, "action", List.of("website"), List.of("OK"), List.of());
+
+        var message = new ActionLogMessage();
+        message.app = "website";
+        message.result = "WARN";
+        forwarder.forward(List.of(message));
+        verifyNoInteractions(publisher);
+    }
+
+    @Test
+    void forwardWithIgnoredErrorCode() {
+        var forwarder = new ActionLogForwarder(publisher, "action", List.of("website"), List.of(), List.of("NOT_FOUND"));
+
+        var message = new ActionLogMessage();
+        message.app = "website";
+        message.result = "WARN";
+        message.errorCode = "NOT_FOUND";
+        forwarder.forward(List.of(message));
+        verifyNoInteractions(publisher);
     }
 }