+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Change log
 
+### 7.10.8 (05/16/2022 - )
+
+* log-processor: action forward supports by result
+  > e.g. only forward OK actions
+
 ### 7.10.7 (04/29/2022 - 05/12/2022)
 
 * log-collector: make http().maxForwardedIPs() read property "sys.http.maxForwardedIPs"